+++ src/Components/Button/styles.css
@@ -9,7 +9,6 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    padding: 0.5vw;
 
     .cta-button-text {
         color: white;
@@ -18,6 +17,7 @@
         font-size: 2rem;
         position: relative;
         z-index: 1;
+        margin: 0.2rem 0.7rem;
 
         @media (max-width: 1365px) {
             font-size: 1.7rem;
