+++ src/Components/CreateSale/styles.css
@@ -1,26 +1,22 @@
 .create-sale-container {
     width: 30%;
     height: 70%;
-    position: fixed;
+    position: absolute;
     align-self: center;
     justify-self: center;
     justify-content: space-around;
     display: flex;
     flex-direction: column;
     align-items: center;
-    z-index: 2;
-    background-color: #1C1D1F;
-    border-color:  #272a2c;
-    box-shadow: 0px 0px 0px 100vw rgba(33, 35, 37, 0.9);
+    z-index: 1;
 
-    @media (min-width: 768px) and (max-width: 1365px) {
-        width: 90%;
-        height: 80%;
+    @media (max-width: 1365px) {
+        width: 40%;
+        height: 20%;
     }
-
     @media (max-width: 767px) {
         width: 90%;
-        height: 80%;
+        height: 10%;
     }
 
 
@@ -32,8 +28,29 @@
         align-self: flex-start;
     }
 
+    .skin-preview {
+        position: absolute;
+        width: 150px;
+        height: 0px;
+    }
+
+    .create-sale-form {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: space-evenly;
+        height: 60%;
+
+
+    }
+
     .close-create-sale:hover {
         color: #F52B00;
+    }
+
+    .price-input {
+        border: 1px solid #212325;
+        background-color: rgba(33, 35, 37, 0.3);
     }
 
     .buttons-container {
