--- src/Components/CreateSaleForm/styles.css
@@ -1,14 +1,0 @@
-.create-sale-form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly;
-    height: 60%;
-
-
-    .skin-preview {
-        position: absolute;
-        width: 150px;
-        height: 0px;
-    }
-}
