--- src/Components/FilterError/styles.css
@@ -1,33 +1,0 @@
-.error-empty-query {
-    width: 45%;
-    height: 50%;
-    background: #1C1D1F;
-    border: 1px solid #212325;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    border-radius: 5px;
-    margin: auto auto;
-    gap: 1rem;
-    grid-column: 1/-1;
-
-    @media (max-width: 1365px) {
-        width: 90%;
-        height: 90%;
-    }
-
-    .caution-icon {
-        width: 40%;
-        height: 40%;
-        color: #FF0000;
-    }
-
-    .error-text {
-        color: #ffffff;
-        font-family: oswald;
-        font-weight: 300;
-        font-size: 1.3rem;
-        text-align: center;
-    }
-}
