--- src/Components/Icons/CautionIcon/index.tsx
@@ -1,24 +1,0 @@
-
-function CautionIcon() {
-
-    return (
-
-        <svg className="caution-icon" width="512" height="512" viewBox="0 0 512 512" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <g clip-path="url(#clip0_17_34)">
-                <path d="M385.08 0H375.688H282.118H229.88H126.92L0 126.929V229.88V282.119V385.069L126.92 512H229.88H282.118H385.08L512 385.07V282.12V229.88V126.929L385.08 0ZM466.68 229.88V282.119V366.296L366.296 466.67H282.118H229.88H145.704L45.34 366.296V282.12V229.88V145.682L145.704 45.33H229.88H282.118H366.296L466.68 145.682V229.88Z" fill="currentColor" />
-                <path d="M284.53 125.07H227.47V292.925H284.53V125.07Z" fill="currentColor" />
-                <path d="M284.53 333.226H227.47V386.94H284.53V333.226Z" fill="currentColor" />
-            </g>
-            <defs>
-                <clipPath id="clip0_17_34">
-                    <rect width="512" height="512" fill="white" />
-                </clipPath>
-            </defs>
-        </svg>
-
-    )
-
-}
-
-
-export default CautionIcon
