+++ src/Components/SalesCategoriesFilter/index.tsx
@@ -8,13 +8,6 @@ import Dropdown from "../Dropdown"
 
 import { salesFiltersContext } from "../../Context/salesFiltersContext"
 
-import useUpdateFilters from "../../Hooks/useUpdateFilters"
-
-import useFetch from "../../Hooks/useFetch"
-
-import useArray from "../../Hooks/useArray"
-
-
 
 const API_SEARCH_CATEGORIES: string = "https://api-skin-warriors.onrender.com/skins/search-categories"
 
@@ -23,15 +16,14 @@ function SalesCategoriesFilter({ updateCreateSaleState }: any) {
 
     const salesFiltersContextValue = useContext(salesFiltersContext)
 
+
     if (!salesFiltersContextValue) throw new Error("usesalesFiltersContext deve ser usado dentro de um salesFiltersProvider")
 
-    const { salesFilters } = salesFiltersContextValue
+
+    const { salesFilters, setSalesFilters } = salesFiltersContextValue
+
 
     const [categories, setCategories] = useState<any>(null)
-
-    const updateCategory = useUpdateFilters()
-
-    const [fetchValue, fetchEndpoint] = useFetch<any>()
 
 
     useEffect(() => {
@@ -40,22 +32,22 @@ function SalesCategoriesFilter({ updateCreateSaleState }: any) {
 
     }, [])
 
-    useEffect(() => {
-
-        transformToArray()
-
-    }, [fetchValue])
-
 
     async function listCategories() {
 
-        await fetchEndpoint(API_SEARCH_CATEGORIES)
+        const response: Response = await fetch(API_SEARCH_CATEGORIES)
+
+        const data: Sale = await response.json()
+
+        transformToArray(data, setCategories)
 
     }
 
-    function transformToArray() {
+    function transformToArray(object: Sale, updateState: any) {
 
-        useArray(fetchValue, setCategories)
+        const array: any[] = Object.values(object)
+
+        updateState(array)
 
     }
 
@@ -67,7 +59,50 @@ function SalesCategoriesFilter({ updateCreateSaleState }: any) {
 
     function updateCategoryState(weapon: string) {
 
-        updateCategory("weapon", weapon)
+        if (salesFilters?.currentWeapon === undefined && salesFilters?.currentWear === undefined) {
+
+            setSalesFilters((prevObject: any) => ({ ...prevObject, currentWeapon: weapon, currentPage: "1" }))
+
+
+        }
+
+        else if (salesFilters?.currentWeapon === undefined && salesFilters?.currentWear !== undefined) {
+
+            setSalesFilters((prevObject: any) => ({ ...prevObject, currentWeapon: weapon, currentPage: "1" }))
+
+
+        }
+
+        else if (salesFilters?.currentWeapon !== undefined && salesFilters?.currentWear === undefined) {
+
+
+            if (weapon !== salesFilters?.currentWeapon) {
+
+                setSalesFilters((prevObject: any) => ({ ...prevObject, currentWeapon: weapon, currentPage: "1" }))
+
+
+            } else {
+
+                setSalesFilters((prevObject: any) => ({ ...prevObject, currentWeapon: undefined, currentPage: "1" }))
+
+
+            }
+
+        }
+
+        else if (salesFilters?.currentWeapon && salesFilters?.currentWear !== undefined) {
+
+            if (weapon !== salesFilters?.currentWeapon) {
+
+                setSalesFilters((prevObject: any) => ({ ...prevObject, currentWeapon: weapon, currentPage: "1" }))
+
+            } else {
+
+                setSalesFilters((prevObject: any) => ({ ...prevObject, currentWeapon: undefined, currentPage: "1" }))
+
+            }
+
+        }
 
     }
