+++ src/Components/SalesCategoriesFilter/styles.css
@@ -28,7 +28,6 @@
 
         @media (max-width: 768px) {
             width: 30%;
-            margin-bottom: 0.5rem;
         }
     }
