--- src/Components/SalesList/styles.css
@@ -1,24 +1,0 @@
-.skins-list-container {
-    grid-area: skinsList;
-    display: grid;
-    grid-template-columns: repeat(7, 1fr);
-    gap: 0.7rem;
-    width: 100%;
-    justify-content: flex-start;
-
-    @media (max-width: 767px) {
-        width: 80%;
-        grid-template-columns: repeat(3, 1fr);
-    }
-
-    @media (min-width: 768px) and (max-width: 1365px) {
-        width: 80%;
-        grid-template-columns: repeat(5, 1fr);
-    }
-
-
-    .skins-list-item {
-        width: 100%;
-
-    }
-}
