+++ src/Components/SkinCard/styles.css
@@ -2,10 +2,21 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    width: 100%;
+    gap: 1.2vw;
+    width: 95%;
+    height: 95%;
     max-width: 200px;
-    min-height: 300px;
+    height: 350px;
     justify-content: space-evenly;
+}
+
+.horizontal-direction{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-evenly;
+    width: 400px;
+    height: 200px;
 }
 
 .skin-card-container {
@@ -18,7 +29,6 @@
         display: flex;
         flex-direction: column;
         align-items: center;
-        position: relative;
 
         .skin-image {
             width: 90%;
@@ -31,10 +41,9 @@
         background-image: url('../../assets/skinCardBg.svg');
         content: ' ';
         position: absolute;
-        width: 100%;
-        height: 100%;
+        width: 80%;
+        height: 80%;
         background-size: contain;
-        background-position: center;
         background-repeat: no-repeat;
         z-index: -1;
     }
@@ -102,4 +111,19 @@
 
 
     }
+}
+
+.skin-infos-horizontal {
+    width: 50%;
+}
+
+.remove-item{
+    width: 10%;
+    height: 10%;
+    color: white;
+    transition: 0.2s ease-out;
+}
+
+.remove-item:hover {
+    color: #F52B00;
 }
