+++ src/Pages/HomeScreen/Home.css
@@ -1,6 +1,8 @@
 .container-home {
     width: 100%;
     height: 90vh;
+    display: flex;
+    flex-direction: column;
     align-items: center;
     justify-content: flex-end;
     background-image: url("../../assets/heroImageDesktop.png");
