+++ src/Pages/HomeScreen/Home.tsx
@@ -2,7 +2,17 @@ import "./Home.css"
 
 import Button from "../../Components/Button/"
 
-import { NavigateFunction, useNavigate } from "react-router-dom"
+import { useNavigate } from "react-router-dom"
+
+
+const styles = {
+
+    container: "container-home",
+
+    button: "home-cta-button"
+
+}
+
 
 function Home() {
 
@@ -10,20 +20,20 @@ function Home() {
 
     const buttonTitle: string = "OFERTAS"
 
-    const navigate: NavigateFunction = useNavigate()
+
+    const navigate = useNavigate()
 
 
-    function changePage(): void {
+    function changePage() {
 
         navigate(ofertasPageURL)
 
     }
 
-
     return (
 
-        <main className="container-column container-home">
-            <Button className="home-cta-button" title={buttonTitle} onClick={changePage} />
+        <main className={styles.container}>
+            <Button className={styles.button} title={buttonTitle} onClick={changePage} />
         </main>
 
     )
