+++ src/Pages/OfertasScreen/OfertasScreen.css
@@ -15,23 +15,125 @@
         display: flex;
         flex-direction: column;
         align-items: center;
-        justify-content: space-between;
     }
 
-    .cta-button-text {
-        color: white;
+    .skins-list-container {
+        grid-area: skinsList;
+        display: flex;
+        flex-direction: row;
+        align-items: flex-start;
+        flex-wrap: wrap;
+        gap: 0.7rem;
+        width: 100%;
+        justify-content: flex-start;
+
+        @media (max-width: 767px) {
+            justify-content: center;
+        }
+
+        @media (min-width: 768px) and (max-width: 1365px) {
+            width: 80%;
+            justify-content: flex-start;
+        }
+
+
+
+        .skins-list-item {
+            width: 13%;
+
+            @media (max-width: 767px) {
+                width: 33%;
+            }
+
+            @media (min-width: 768px) and (max-width: 1365px) {
+                width: 23%;
+            }
+
+        }
+    }
+}
+
+.cta-button-text {
+    color: white;
+    font-family: oswald;
+    font-weight: 300;
+    font-size: 2rem;
+    position: relative;
+    z-index: 1;
+}
+
+.skeleton-load-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+    max-width: 200px;
+    min-height: 300px;
+    justify-content: space-evenly;
+
+
+    .skeleton-load-image {
+        width: 90%;
+        height: 120px;
+        border-radius: 5px;
+        animation: skeletonLoad 2s linear infinite alternate;
+    }
+
+    .skeleton-load-text {
+        width: 120px;
+        height: 20px;
+        animation: skeletonLoad 2s linear infinite alternate;
+    }
+
+}
+
+.enabled-weapon {
+    color: #F52B00 !important;
+}
+
+.disabled-weapon {
+    color: #ffffff !important;
+}
+
+.error-empty-query {
+    width: 45%;
+    height: 50%;
+    background: #1C1D1F;
+    border: 1px solid #212325;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    border-radius: 5px;
+    margin: auto auto;
+    gap: 1rem;
+
+    @media (max-width: 768px) {
+        width: 80%;
+    }
+
+    .caution-icon {
+        width: 40%;
+        height: 40%;
+        color: #FF0000;
+    }
+
+    .error-text {
+        color: #ffffff;
         font-family: oswald;
         font-weight: 300;
-        font-size: 2rem;
-        position: relative;
-        z-index: 1;
+        font-size: 1.3rem;
+        text-align: center;
+    }
+}
+
+@keyframes skeletonLoad {
+
+    0% {
+        background-color: hsl(210, 6%, 14%);
     }
 
-    .enabled-weapon {
-        color: #F52B00 !important;
-    }
-
-    .disabled-weapon {
-        color: #ffffff !important;
+    100% {
+        background-color: hsl(210, 6%, 24%);
     }
 }
